--- .env
@@ -1,2 +1,0 @@
-JWT_SECRET=mykey1234
-JWT_EXPIRATION=1m  # You can change this to any duration like '30m', '1d', etc.
