+++ BACKEND_APP/.env
@@ -1,0 +1,3 @@
+JWT_SECRET=mykey1234
+JWT_EXPIRATION=1m  # You can change this to any duration like '30m', '1d', etc.
+CORS_ORIGINS=http://localhost:3000,https://hamrorealstate.store
