+++ BACKEND_APP/.gitignore
@@ -1,0 +1,8 @@
+# Ignore node_modules
+node_modules/
+
+# Ignore logs
+logs/
+
+# Ignore Docker ignore file (optional)
+.dockerignore
