+++ BACKEND_APP/controllers/authController.js
