+++ BACKEND_APP/middlewares/authMiddleware.js
