+++ BACKEND_APP/routes/authRoutes.js
